services/requirements.txt
pymilvus==2.5.0
sentence-transformers==3.1.1
PyYAML==6.0.2
kafka-python==2.0.2
